Created basic upload script
Basic uploading is working.
Uploaded files are stored in the 'upload/' folder.
File names are stripped of special characters to avoid errors when
displaying the video (having a quote character in the filename will
cause HTML to end the string at the wrong place).
Added a requirement for uploading to work in the README.